Add class to HTML for syntax highlighting
Added `class="language-${languageName}"` so that syntax highlighting will hopefully work. Adding a link to the language's website would be nice, but I don't know how to do that or if that would be easy.

Btw, I feel like the `Copied to clipboard!` notification should be sent after it's actually copied, but that's a PR for another day.